Refactor import paths; replace incorrect 'commom' directory references with 'uid' in multiple files for consistency and clarity.
