+++ lib/core/domain/company/company.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 
-import '../commom/uid.dart';
+import '../uid.dart';
 
 class Company extends Equatable {
   final Uid id;
