+++ lib/infra/services/http/tractian_http_response.dart
@@ -1,9 +1,9 @@
-import '../../../core/domain/commom/uid.dart';
 import '../../../core/domain/company/company.dart';
 import '../../../core/domain/company_asset/company_asset.dart';
 import '../../../core/domain/company_asset/sensor_types.dart';
 import '../../../core/domain/company_asset/statuses.dart';
 import '../../../core/domain/company_location/company_location.dart';
+import '../../../core/domain/uid.dart';
 import '../utils.dart';
 
 class TractianHttpResponse {
